deleted unused files, moved some work in progress file in the folder, fix connection issue between components
