+++ app/components/Contact/styles.scss
@@ -1,4 +1,4 @@
-@import '../../constants.scss';
+@import '../../stylesheets/constants.scss';
 
 .contact {
   display: flex;
