+++ app/components/FirstLaunch/styles.scss
