+++ app/components/Occupied/styles.scss
